test: trim the pack index tests to their distinct layers
packindex_migration_test.go was written to validate the sealing change, and
most of it has since been superseded by the v1.14.0 fixture, which pins the
same behaviour against real historical bytes rather than a simulation.

Its migration test built a state no client produces — a plaintext catalog with
sealed footers — since anything that seals footers also seals the catalog. The
real pre-sealing state is now covered end to end in
e2e/feature_legacy_repo_test.go, so that test is removed.

What it uniquely covered is kept: it was the only test exercising
WithPackfile(true) through the public Client, and client_test.go has no
backup/restore round trip at all. The e2e suite cannot stand in, because it
drives the built binary rather than the exported API a library user calls. That
round trip survives as client_packfile_test.go, framed for what it guards.

The one assertion the migration test made that the fixture did not — that an
upgraded repository's catalog is re-sealed on the next flush — moves into the
fixture's backup subtest, so nothing is lost.

Refs #295
